Renamed from "MainActivity". Deletes unused imports. Added comments. @receivedBoolean Simplified "if"
